requirements.txt
aiogram~=3.20.0
python-dotenv~=1.0.0
python-docx~=1.1.0
num2words~=0.5.13